+++ R/rscript.R
@@ -12,8 +12,7 @@
 #' @export
 
 rscript <- function(script, cmdargs = character(), libpath = .libPaths(),
-                    repos = c(getOption("repos"),
-                              c(CRAN = "https://cloud.r-project.org")),
+                    repos = default_repos(),
                     stdout = NULL, stderr = NULL,
                     poll_connection = TRUE, echo = FALSE, show = TRUE,
                     callback = NULL, block_callback = NULL, spinner = FALSE,