+++ tests/testthat/test-utils.R
@@ -20,3 +20,42 @@ test_that("is_complete_expression", {
     withr::with_envvar(c(LANGUAGE = "de_DE"), do_tests())
   }
 })
+
+test_that("default_repos", {
+  def <- "https://cloud.r-project.org"
+
+  withr::with_options(list(repos = NULL),
+    expect_equal(
+      default_repos(),
+      c(CRAN = def)))
+
+  withr::with_options(list(repos = character()),
+    expect_equal(
+      default_repos(),
+      c(CRAN = def)))
+
+  withr::with_options(list(repos = list()),
+    expect_equal(
+      default_repos(),
+      list(CRAN = def)))
+
+  withr::with_options(list(repos = c(foo = "bar")),
+    expect_equal(
+      default_repos(),
+      c(foo = "bar", CRAN = def)))
+
+  withr::with_options(list(repos = list(foo = "bar")),
+    expect_equal(
+      default_repos(),
+      list(foo = "bar", CRAN = def)))
+
+  withr::with_options(list(repos = c(foo = "bar", CRAN = "set")),
+    expect_equal(
+      default_repos(),
+      c(foo = "bar", CRAN = "set")))
+
+  withr::with_options(list(repos = c(foo = "bar", CRAN = "@CRAN@")),
+    expect_equal(
+      default_repos(),
+      c(foo = "bar", CRAN = def)))
+})